+++ FORZE/Support/FZIO.cpp
@@ -56,7 +56,7 @@ namespace FORZE {
         
         
         fseek(f, 0, SEEK_END);
-        size_t size = ftell(f);
+        size_t size = (size_t)ftell(f);
         fseek(f, 0, SEEK_SET);
